requirements.txt
# Dev-time
pipdeptree
pytest

# Runtime
pyyaml